refactor(types): move IComponent to components/index.d.ts
